WIP: Refactor matching, using syntax-table, etc.
Following some of rainbow-blocks's example, we copy the syntax table
and also match [] and {}.  The prism-match function is rewritten using
scan-lists, and this seems simpler and easier to understand.

Note that this currently colorizes strings like everything else.  I
think that is a nice feature, or would be if it colorized them
slightly differently, so let's see...