WIP: Debugging region extension
It does seem to be necessary in some cases.  Maybe refactoring of the
matcher could avoid it, but I'm not sure.  It's getting complicated.

A problem arises when typing into a form whose containing list spans
large portions of the buffer.  The region can be extended to
potentially tens of thousands of characters beyond the point.  Putting
those messages in the extension function showed that after each
extension, it was being called again starting one character after the
previous extension, which I don't understand, and that was happening
until almost the end of the buffer.  It seemed to be exiting more than
the necessary lists as well.

It's probably not going to be a problem in many Elisp files.  In my
case, it was happening in a (use-package org ...) form, which contains
a lot of code that's deeper than depth 0.

I hope it can be fixed, but it will take a lot of experimentation, I
think.